requirements.txt
pdfplumber
pydantic
python-dateutil
openpyxl
fastapi
uvicorn
jinja2
python-multipart
